fix(US-004): split oversized test file and remove lint violations
- Split runner-parallel-metrics.test.ts (635 lines) into two files
  under the 400-line hard limit:
  runner-parallel-metrics-ac1-ac2.test.ts (AC-1, AC-2, AC-5) — 341 lines
  runner-parallel-metrics-ac3-ac4.test.ts (AC-3, AC-4) — 369 lines
- Remove unused _parallelBatchDeps destructure in AC-2 type stub test
- Replace Node.js fs APIs (existsSync, mkdtempSync, rmSync) with
  Bun-native lifecycle: initLogger/resetLogger + mock.restore()
  (the tmpDir was created but never used in any test body)
